requirements.txt
Flask
Flask-SocketIO
transformers
torch
peft
datasets
PyYAML
matplotlib
seaborn
networkx
pandas
psutil
watchdog
gunicorn
redis
Flask==2.3.2
Flask-SocketIO==5.3.4
transformers==4.30.2
torch==2.0.1+cpu
peft==0.4.0
datasets==2.14.0
PyYAML==6.0
matplotlib==3.7.2
seaborn==0.12.2
networkx==3.1
pandas==2.0.3
psutil==5.9.5
watchdog==3.0.0
gunicorn==20.1.0